app/requirements.txt
line-bot-sdk
requests
beautifulsoup4
